Refractoring Code:
deleteTodo: Has been removed from entirely. Was able to, with the help
of Evan, to move the functionality to emit listen on the component tag.
@delete-todo='todo.splice(index, 1)'.

addTodo: Added a randomly generated ID number for the todo items, so
they are unique. New function generateID was created as a helper.

CSS: Minor updates